Add react-dom as a peerDependency (#66)
This appears to resolve a problem where active visits seems to be using
a version of react-dom that is bundled in the carbon-components-react
package. The problem causes React Dev Tools to see the second react-dom
as a new renderer, which can cause it to erroneously detect the page
as using the production build of React when using a local development
server.